requirements.txt
absl-py==0.1.10
astor==0.6.2
backports.weakref==1.0.post1
bleach==1.5.0
enum34==1.1.6
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
imutils==0.4.5
Keras==2.1.3
Markdown==2.6.11
mock==2.0.0
numexpr==2.6.4
numpy==1.14.1
opencv-contrib-python==3.4.0.12
pbr==3.1.1
progressbar2==3.36.0
protobuf==3.5.1
python-utils==2.3.0
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tables==3.4.2
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
Werkzeug==0.14.1
